+++ packages/angular-query-experimental/src/query-options.ts
@@ -53,7 +53,7 @@ export type DefinedInitialDataOptions<
 }
 
 /**
- * Allows to share and re-use query options in a type-safe way.
+ * Allows sharing and re-using query options in a type-safe way.
  *
  * The `queryKey` will be tagged with the type from `queryFn`.
  *
@@ -84,7 +84,7 @@ export function queryOptions<
   QueryKeyWithDataTag<TQueryKey, TQueryFnData, TError>
 
 /**
- * Allows to share and re-use query options in a type-safe way.
+ * Allows sharing and re-using query options in a type-safe way.
  *
  * The `queryKey` will be tagged with the type from `queryFn`.
  *
@@ -115,7 +115,7 @@ export function queryOptions<
   QueryKeyWithDataTag<TQueryKey, TQueryFnData, TError>
 
 /**
- * Allows to share and re-use query options in a type-safe way.
+ * Allows sharing and re-using query options in a type-safe way.
  *
  * The `queryKey` will be tagged with the type from `queryFn`.
  *
@@ -146,7 +146,7 @@ export function queryOptions<
   QueryKeyWithDataTag<TQueryKey, TQueryFnData, TError>
 
 /**
- * Allows to share and re-use query options in a type-safe way.
+ * Allows sharing and re-using query options in a type-safe way.
  *
  * The `queryKey` will be tagged with the type from `queryFn`.
  *