+++ packages/preact-query/src/__tests__/QueryResetErrorBoundary.test.tsx
@@ -27,8 +27,8 @@ describe('QueryErrorResetBoundary', () => {
   })
 
   afterEach(() => {
-    vi.useRealTimers()
     queryClient.clear()
+    vi.useRealTimers()
   })
 
   describe('useQuery', () => {