+++ packages/preact-query/src/__tests__/usePrefetchQuery.test.tsx
@@ -11,14 +11,8 @@ import {
   useQueryErrorResetBoundary,
   useSuspenseQuery,
 } from '..'
-import type { UseSuspenseQueryOptions } from '..'
 import { ErrorBoundary } from './ErrorBoundary'
 import { renderWithClient } from './utils'
-
-const generateQueryFn = (data: string) =>
-  vi
-    .fn<(...args: Array<any>) => Promise<string>>()
-    .mockImplementation(() => sleep(10).then(() => data))
 
 describe('usePrefetchQuery', () => {
   let queryCache: QueryCache
@@ -31,33 +25,24 @@ describe('usePrefetchQuery', () => {
   })
 
   afterEach(() => {
-    vi.useRealTimers()
     queryClient.clear()
+    vi.useRealTimers()
   })
-
-  function Suspended<TData = unknown>(props: {
-    queryOpts: UseSuspenseQueryOptions<TData, Error, TData, Array<string>>
-    children?: VNode
-  }) {
-    const state = useSuspenseQuery(props.queryOpts)
-
-    return (
-      <div>
-        <div>data: {String(state.data)}</div>
-        {props.children}
-      </div>
-    )
-  }
 
   it('should prefetch query if query state does not exist', async () => {
     const queryOpts = {
       queryKey: queryKey(),
-      queryFn: generateQueryFn('prefetchQuery'),
+      queryFn: vi.fn(() => sleep(10).then(() => 'prefetchQuery')),
     }
 
     const componentQueryOpts = {
       ...queryOpts,
-      queryFn: generateQueryFn('useSuspenseQuery'),
+      queryFn: () => sleep(10).then(() => 'useSuspenseQuery'),
+    }
+
+    function Page() {
+      const state = useSuspenseQuery(componentQueryOpts)
+      return <div>data: {String(state.data)}</div>
     }
 
     function App() {
@@ -65,12 +50,14 @@ describe('usePrefetchQuery', () => {
 
       return (
         <Suspense fallback="Loading...">
-          <Suspended queryOpts={componentQueryOpts} />
+          <Page />
         </Suspense>
       )
     }
 
     const rendered = renderWithClient(queryClient, <App />)
+
+    expect(rendered.getByText('Loading...')).toBeInTheDocument()
 
     await vi.advanceTimersByTimeAsync(10)
     expect(rendered.getByText('data: prefetchQuery')).toBeInTheDocument()
@@ -80,7 +67,14 @@ describe('usePrefetchQuery', () => {
   it('should not prefetch query if query state exists', async () => {
     const queryOpts = {
       queryKey: queryKey(),
-      queryFn: generateQueryFn('The usePrefetchQuery hook is smart!'),
+      queryFn: vi.fn(() =>
+        sleep(10).then(() => 'The usePrefetchQuery hook is smart!'),
+      ),
+    }
+
+    function Page() {
+      const state = useSuspenseQuery(queryOpts)
+      return <div>data: {String(state.data)}</div>
     }
 
     function App() {
@@ -88,7 +82,7 @@ describe('usePrefetchQuery', () => {
 
       return (
         <Suspense fallback="Loading...">
-          <Suspended queryOpts={queryOpts} />
+          <Page />
         </Suspense>
       )
     }
@@ -108,18 +102,23 @@ describe('usePrefetchQuery', () => {
   it('should let errors fall through and not refetch failed queries', async () => {
     const consoleMock = vi.spyOn(console, 'error')
     consoleMock.mockImplementation(() => undefined)
-    const queryFn = generateQueryFn('Not an error')
+    const queryFn = vi.fn(() => sleep(10).then(() => 'Not an error'))
 
     const queryOpts = {
       queryKey: queryKey(),
       queryFn,
     }
 
-    queryFn.mockImplementationOnce(async () => {
-      await sleep(10)
+    queryFn.mockImplementationOnce(() =>
+      sleep(10).then(() => {
+        throw new Error('Oops! Server error!')
+      }),
+    )
 
-      throw new Error('Oops! Server error!')
-    })
+    function Page() {
+      const state = useSuspenseQuery(queryOpts)
+      return <div>data: {String(state.data)}</div>
+    }
 
     function App() {
       usePrefetchQuery(queryOpts)
@@ -127,7 +126,7 @@ describe('usePrefetchQuery', () => {
       return (
         <ErrorBoundary fallbackRender={() => <div>Oops!</div>}>
           <Suspense fallback="Loading...">
-            <Suspended queryOpts={queryOpts} />
+            <Page />
           </Suspense>
         </ErrorBoundary>
       )
@@ -146,7 +145,7 @@ describe('usePrefetchQuery', () => {
   })
 
   it('should not create an endless loop when using inside a suspense boundary', async () => {
-    const queryFn = generateQueryFn('prefetchedQuery')
+    const queryFn = vi.fn(() => sleep(10).then(() => 'prefetchedQuery'))
 
     const queryOpts = {
       queryKey: queryKey(),
@@ -158,11 +157,16 @@ describe('usePrefetchQuery', () => {
       return <>{children}</>
     }
 
+    function Page() {
+      const state = useSuspenseQuery(queryOpts)
+      return <div>data: {String(state.data)}</div>
+    }
+
     function App() {
       return (
         <Suspense fallback={<></>}>
           <Prefetch>
-            <Suspended queryOpts={queryOpts} />
+            <Page />
           </Prefetch>
         </Suspense>
       )
@@ -177,18 +181,25 @@ describe('usePrefetchQuery', () => {
   it('should be able to recover from errors and try fetching again', async () => {
     const consoleMock = vi.spyOn(console, 'error')
     consoleMock.mockImplementation(() => undefined)
-    const queryFn = generateQueryFn('This is fine :dog: :fire:')
+    const queryFn = vi.fn(() =>
+      sleep(10).then(() => 'This is fine :dog: :fire:'),
+    )
 
     const queryOpts = {
       queryKey: queryKey(),
       queryFn,
     }
 
-    queryFn.mockImplementationOnce(async () => {
-      await sleep(10)
+    queryFn.mockImplementationOnce(() =>
+      sleep(10).then(() => {
+        throw new Error('Oops! Server error!')
+      }),
+    )
 
-      throw new Error('Oops! Server error!')
-    })
+    function Page() {
+      const state = useSuspenseQuery(queryOpts)
+      return <div>data: {String(state.data)}</div>
+    }
 
     function App() {
       const { reset } = useQueryErrorResetBoundary()
@@ -205,7 +216,7 @@ describe('usePrefetchQuery', () => {
           )}
         >
           <Suspense fallback="Loading...">
-            <Suspended queryOpts={queryOpts} />
+            <Page />
           </Suspense>
         </ErrorBoundary>
       )
@@ -230,20 +241,47 @@ describe('usePrefetchQuery', () => {
   it('should not create a suspense waterfall if prefetch is fired', async () => {
     const firstQueryOpts = {
       queryKey: queryKey(),
-      queryFn: generateQueryFn('Prefetch is nice!'),
+      queryFn: vi.fn(() => sleep(10).then(() => 'Prefetch is nice!')),
     }
 
     const secondQueryOpts = {
       queryKey: queryKey(),
-      queryFn: generateQueryFn('Prefetch is really nice!!'),
+      queryFn: vi.fn(() => sleep(10).then(() => 'Prefetch is really nice!!')),
     }
 
     const thirdQueryOpts = {
       queryKey: queryKey(),
-      queryFn: generateQueryFn('Prefetch does not create waterfalls!!'),
+      queryFn: vi.fn(() =>
+        sleep(10).then(() => 'Prefetch does not create waterfalls!!'),
+      ),
     }
 
     const Fallback = vi.fn().mockImplementation(() => <div>Loading...</div>)
+
+    function FirstQuery({ children }: { children?: VNode }) {
+      const state = useSuspenseQuery(firstQueryOpts)
+      return (
+        <div>
+          <div>data: {String(state.data)}</div>
+          {children}
+        </div>
+      )
+    }
+
+    function SecondQuery({ children }: { children?: VNode }) {
+      const state = useSuspenseQuery(secondQueryOpts)
+      return (
+        <div>
+          <div>data: {String(state.data)}</div>
+          {children}
+        </div>
+      )
+    }
+
+    function ThirdQuery() {
+      const state = useSuspenseQuery(thirdQueryOpts)
+      return <div>data: {String(state.data)}</div>
+    }
 
     function App() {
       usePrefetchQuery(firstQueryOpts)
@@ -252,11 +290,11 @@ describe('usePrefetchQuery', () => {
 
       return (
         <Suspense fallback={<Fallback />}>
-          <Suspended queryOpts={firstQueryOpts}>
-            <Suspended queryOpts={secondQueryOpts}>
-              <Suspended queryOpts={thirdQueryOpts} />
-            </Suspended>
-          </Suspended>
+          <FirstQuery>
+            <SecondQuery>
+              <ThirdQuery />
+            </SecondQuery>
+          </FirstQuery>
         </Suspense>
       )
     }